Authorize 1% of pixels error.

+++ pkg/test/test.go
@@ -29,13 +29,17 @@ func VerifyImage(t *testing.T, resultPath string) {
 	masterBound, masterPixels := get(t, masterPath)
 	resultBound, resultPixels := get(t, resultPath)
 
+	assert.Len(t, resultPixels, len(masterPixels))
 	if len(resultPixels) == 0 || len(masterPixels) == 0 {
 		return
 	}
-
-	assert.Len(t, resultPixels, len(masterPixels))
 	assert.Equal(t, masterBound, resultBound)
-	assert.Equal(t, masterPixels, resultPixels)
+	if masterBound != resultBound {
+		return
+	}
+
+	ratio := compare(masterPixels, resultPixels)
+	assert.LessOrEqual(t, ratio, 1)
 
 	if !t.Failed() {
 		os.Remove(resultPath)
@@ -72,4 +76,21 @@ func pixels(i image.Image) ([]uint8, error) {
 	default:
 		return nil, fmt.Errorf("unsupported image type %T", i)
 	}
+}
+
+func compare(img1, img2 []uint8) int {
+	overallChange := 0
+
+	for i := 0; i < len(img1); i++ {
+		overallChange += diffUInt8(img1[i], img2[i])
+	}
+
+	return 100 * overallChange / len(img1)
+}
+
+func diffUInt8(x, y uint8) int {
+	if x != y {
+		return 1
+	}
+	return 0
 }
